+++ apps/graphql-api/src/models/Product.ts
