+++ apps/graphql-api/src/models/ProductAdditionalInfo.ts
