+++ apps/graphql-api/src/models/ProductImage.ts
