+++ apps/graphql-api/src/schema/schema.ts
@@ -1,6 +1,6 @@
-import Product from "../data/Product";
-import ProductAdditionalInfo from "../data/ProductAdditionalInfo";
-import ProductImage from "../data/ProductImage";
+import Product from "../models/Product";
+import ProductAdditionalInfo from "../models/ProductAdditionalInfo";
+import ProductImage from "../models/ProductImage";
 import ProductType from "./types/ProductType";
 import ProductAdditionalInfoType from "./types/ProductAdditionalInfoType";
 import ProductImageType from "./types/ProductImageType";
